Changed app.py and tester.py

+++ src/main/python/LM/tester.py
@@ -1,4 +1,3 @@
-import tokenizers
 from transformers import GPT2Tokenizer, GPTNeoForCausalLM, pipeline, TrainingArguments, Trainer
 from datasets import load_dataset, load_metric
 import random
@@ -12,8 +11,6 @@ import numpy as np
 import os
 import os.path
 
-metric = load_metric("accuracy")
-
 
 class ModelData:
     model = None
@@ -24,79 +21,94 @@ class ModelData:
     teamName = "EleutherAI/"
     modelName = "gpt-neo-125M"
     model_save_directory = "./" + modelName
+    fine_tuned_model_directory = model_save_directory + "_fine-tuned"
     device = 0
     hasInitialized = False
     enable_model = True
     is_locked = False
-    num_runs = 1
+    num_runs = 50
     temperature = 0.9
     dataset = None
+    training_dataset = None
+    validation_dataset = None
 
 
 def tokenize_function(examples):
     inputs = [""] * len(examples["output"])
     for i in range(len(examples["output"])):
         inputs[i] = random.choice(examples["inputs"][i]["input"])
-    return ModelData.tokenizer(examples["output"], inputs, padding="max_length", truncation=True)
+    return ModelData.tokenizer(inputs, examples["output"], padding="max_length", truncation=True)
 
 
 def tokenize_training_function(examples):
     inputs = [""] * len(examples["output"])
     for i in range(len(examples["output"])):
         inputs[i] = random.choice(examples["inputs"][i]["input"])
+        # inputs[i] = examples["inputs"][i]["input"]
     return ModelData.tokenizer(inputs, padding="max_length", truncation=True)
 
 
 def tokenize_valid_function(examples):
-    inputs = [""] * len(examples["output"])
-    for i in range(len(examples["output"])):
-        inputs[i] = random.choice(examples["inputs"][i]["input"])
     return ModelData.tokenizer(examples["output"], padding="max_length", truncation=True)
 
 
-def compute_metrics(eval_pred):
-    logits, labels = eval_pred
-    predictions = np.argmax(logits, axis=-1)
-    return metric.compute(predictions=predictions, references=labels)
+def concatenate(examples):
+    block_size = 128
+    concatenated_examples = {k: sum(examples[k], []) for k in examples.keys()}
+    total_length = len(concatenated_examples[list(examples.keys())[0]])
+    total_length = (total_length // block_size) * block_size
+    result = {
+        k: [t[i: i + block_size] for i in range(0, total_length, block_size)]
+        for k, t in concatenated_examples.items()
+    }
+    examples["labels"] = examples["ouput_id"].copy()
+    return result
 
 
 def train():
-    tokenized_training_datasets = ModelData.dataset.map(tokenize_training_function, batched=True)
-    tokenized_valid_datasets = ModelData.dataset.map(tokenize_valid_function, batched=True)
-    tokenized_training_datasets = tokenized_training_datasets.remove_columns(["id", "output", "inputs"])
-    tokenized_valid_datasets = tokenized_valid_datasets.remove_columns(["id", "output", "inputs"])
-    tokenized_training_datasets = tokenized_training_datasets.with_format("torch")
-    tokenized_valid_datasets = tokenized_valid_datasets.with_format("torch")
+    if not ModelData.validation_dataset:
+        tokenized_valid_datasets = ModelData.dataset.map(tokenize_valid_function, batched=True)
+        tokenized_training_datasets = ModelData.dataset.map(tokenize_training_function, batched=True)
+        # lm_valid_datasets = tokenized_valid_datasets.map(
+        #     concatenate,
+        #     batched=True,
+        #     batch_size=1000
+        # )
+
+        ModelData.tokenizer.decode(tokenized_valid_datasets["train"][1]["input_ids"])
+        lm_valid_datasets = tokenized_valid_datasets.with_format("torch")
+        ModelData.validation_dataset = lm_valid_datasets
+
+    tokenized_training_datasets = ModelData.dataset.map(tokenize_training_function, batched=True,
+                                                        remove_columns=["id", "inputs"])
+    # lm_train_datasets = tokenized_training_datasets.map(
+    #     concatenate,
+    #     batched=True,
+    #     batch_size=1000
+    # )
+    ModelData.tokenizer.decode(tokenized_training_datasets["train"][1]["input_ids"])
+    lm_train_datasets = tokenized_training_datasets.with_format("torch")
+    ModelData.training_dataset = tokenized_training_datasets
+    tokenized_training_datasets["train"]["labels"] = tokenized_valid_datasets["train"]["input_id"].copy()
     args = TrainingArguments(
         output_dir="test_trainer-ds",
-        per_device_train_batch_size=32,
-        per_device_eval_batch_size=32,
-        evaluation_strategy="steps",
-        eval_steps=5_000,
-        logging_steps=5_000,
-        gradient_accumulation_steps=8,
-        num_train_epochs=1,
-        weight_decay=0.1,
-        warmup_steps=1_000,
-        lr_scheduler_type="cosine",
-        learning_rate=5e-4,
-        save_steps=5_000,
-        fp16=True,
+        per_device_train_batch_size=8,
+        per_device_eval_batch_size=8,
+        evaluation_strategy="epoch",
+        learning_rate=2e-5,
+        weight_decay=0.01,
         push_to_hub=False,
     )
     torch.cuda.empty_cache()
     trainer = Trainer(
         model=ModelData.model,
         args=args,
-        train_dataset=tokenized_training_datasets["train"],
-        eval_dataset=tokenized_valid_datasets["train"]
+        train_dataset=lm_train_datasets["train"],
+        eval_dataset=ModelData.validation_dataset["train"]
     )
     trainer.train()
+    trainer.save_model(ModelData.model_save_directory + "_fine-tuned")
 
-    # ModelData.dataset = ModelData.dataset.map(lambda e: ModelData.tokenizer(e['output'], truncation=True), batched=True)
-    # ModelData.dataset.set_format(type='torch', columns=['id', 'output', 'inputs'])
-    # dataloader = torch.utils.data.DataLoader(ModelData.dataset, batch_size=32)
-    # next(iter(dataloader))
     return "Success"
 
 
@@ -110,23 +122,32 @@ def init():
     print("Searching for model...", end=" ")
 
     # Model not present on disk. Download and save.
-    if not os.path.isfile(ModelData.model_save_directory + "/pytorch_model.bin"):
-        print(f"Failed. Model not found, proceeding to download.\n")
-        print("Downloading model...", end=" ")
+    if not os.path.isfile(ModelData.fine_tuned_model_directory + "/pytorch_model.bin"):
+        if not os.path.isfile(ModelData.model_save_directory + "/pytorch_model.bin"):
+            print(f"Failed. Model not found, proceeding to download.\n")
+            print("Downloading model...", end=" ")
+            start_time = time.perf_counter()
+            GPTNeoForCausalLM.from_pretrained(ModelData.teamName + ModelData.modelName).save_pretrained(
+                ModelData.model_save_directory)
+            end_time = time.perf_counter()
+            print(f"Completed after {end_time - start_time:0.2f} sec.\n")
+        # Model is present on disk, don't download!
+        else:
+            print("Found! ")
+
+        # Load model into RAM
+        print("Loading model...", end=" ")
         start_time = time.perf_counter()
-        GPTNeoForCausalLM.from_pretrained(ModelData.teamName + ModelData.modelName).save_pretrained(
-            ModelData.model_save_directory)
+        ModelData.model = GPTNeoForCausalLM.from_pretrained(ModelData.model_save_directory)
         end_time = time.perf_counter()
         print(f"Completed after {end_time - start_time:0.2f} sec.\n")
-
-    # Model is present on disk, don't download!
     else:
-        print("Found! ")
+        print("Found fine-tuned version! ")
 
     # Load model into RAM
     print("Loading model...", end=" ")
     start_time = time.perf_counter()
-    ModelData.model = GPTNeoForCausalLM.from_pretrained(ModelData.model_save_directory)
+    ModelData.model = GPTNeoForCausalLM.from_pretrained(ModelData.fine_tuned_model_directory)
     end_time = time.perf_counter()
     print(f"Completed after {end_time - start_time:0.2f} sec.\n")
 
@@ -186,5 +207,15 @@ def init():
 
 if ModelData.enable_model:
     init()
-    train()
+    for i in range(ModelData.num_runs):
+        if i > 0:
+            print("Loading latest training result!\n")
+            # Load model into RAM
+            print("Loading model...", end=" ")
+            start_time = time.perf_counter()
+            ModelData.model = GPTNeoForCausalLM.from_pretrained(ModelData.fine_tuned_model_directory)
+            end_time = time.perf_counter()
+            print(f"Completed after {end_time - start_time:0.2f} sec.\n")
+        print("Training! Iteration number: " + str(i + 1))
+        train()
